duplicated searchParam for all entities, added auto incrementing ids, added IdGenerator
