+++ src/main/java/com/keyin/entity/city/CitySearchParameters.java
@@ -1,4 +1,4 @@
-package com.keyin.city;
+package com.keyin.entity.city;
 
 public class CitySearchParameters {
     private Long id;
