+++ src/main/java/com/keyin/utility/IdGenerator.java
@@ -1,0 +1,11 @@
+package com.keyin.utility;
+
+public class IdGenerator {
+    private long nextId = 0;
+
+    public long getNextId() {
+        long thisId = nextId;
+        nextId++;
+        return thisId;
+    }
+}
